feat: Add new components for TargetList, ShieldEffect, LaserBeam, and ShootingSystem
- Implemented TargetList component to render a list of targets with hit detection.
- Created ShieldEffect component to visualize shield activation with dynamic scaling and rotation.
- Developed LaserBeam component to render laser beams based on weapon type and properties.
- Introduced ShootingSystem component to manage shooting mechanics, ammo, cooldowns, and explosions.
- Refactored GameCanvas and Player components to utilize new implementations from the asteroid library.
- Updated AmmoIndicator and WeaponDisplay components for better handling of weapon states and ammo.
- Removed old implementations of various components and replaced them with new modular versions.
